+++ tests/unit/study/test_status.py
@@ -30,17 +30,467 @@
 """
 Tests for the Status class in the status.py module
 """
+import json
+import os
 import unittest
 from argparse import Namespace
 from copy import deepcopy
 from datetime import datetime
+from json.decoder import JSONDecodeError
 
+import pytest
 import yaml
 from deepdiff import DeepDiff
+from filelock import Timeout
 
 from merlin.spec.expansion import get_spec_with_expansion
-from merlin.study.status import Status
+from merlin.study.status import Status, read_status, status_conflict_handler, write_status
 from tests.unit.study.status_test_files import shared_tests, status_test_variables
+
+
+class TestStatusReading:
+    """Test the logic for reading in status files"""
+
+    cancel_step_dir = f"{status_test_variables.VALID_WORKSPACE_PATH}/cancel_step"
+    status_file = f"{cancel_step_dir}/MERLIN_STATUS.json"
+    lock_file = f"{cancel_step_dir}/status.lock"
+
+    def test_basic_read(self):
+        """
+        Test the basic reading functionality of `read_status`. There should
+        be no errors thrown and the correct status dict should be returned.
+        """
+        actual_statuses = read_status(self.status_file, self.lock_file)
+        read_status_diff = DeepDiff(
+            actual_statuses, status_test_variables.REQUESTED_STATUSES_JUST_CANCELLED_STEP, ignore_order=True
+        )
+        assert read_status_diff == {}
+
+    def test_timeout_raise_errors_disabled(self, mocker: "Fixture", caplog: "Fixture"):  # noqa: F821
+        """
+        Test the timeout functionality of the `read_status` function with
+        `raise_errors` set to False. This should log a warning message and
+        return an empty dict.
+        This test will create a mock of the FileLock object in order to
+        force a timeout to be raised.
+
+        :param mocker: A built-in fixture from the pytest-mock library to create a Mock object
+        :param caplog: A built-in fixture from the pytest library to capture logs
+        """
+
+        # Set the mock to raise a timeout
+        mock_filelock = mocker.patch("merlin.study.status.FileLock")
+        mock_lock = mocker.MagicMock()
+        mock_lock.acquire.side_effect = Timeout(self.lock_file)
+        mock_filelock.return_value = mock_lock
+
+        # Check that the return is as we expect
+        actual_status = read_status(self.status_file, self.lock_file)
+        assert actual_status == {}
+
+        # Check that a warning is logged
+        expected_log = f"Timed out when trying to read status from '{self.status_file}'"
+        assert expected_log in caplog.text, "Missing expected log message"
+
+    def test_timeout_raise_errors_enabled(self, mocker: "Fixture", caplog: "Fixture"):  # noqa: F821
+        """
+        Test the timeout functionality of the `read_status` function with
+        `raise_errors` set to True. This should log a warning message and
+        raise a Timeout exception.
+        This test will create a mock of the FileLock object in order to
+        force a timeout to be raised.
+
+        :param mocker: A built-in fixture from the pytest-mock library to create a Mock object
+        :param caplog: A built-in fixture from the pytest library to capture logs
+        """
+
+        # Set the mock to raise a timeout
+        mock_filelock = mocker.patch("merlin.study.status.FileLock")
+        mock_lock = mocker.MagicMock()
+        mock_lock.acquire.side_effect = Timeout(self.lock_file)
+        mock_filelock.return_value = mock_lock
+
+        # Check that a Timeout exception is raised
+        with pytest.raises(Timeout):
+            read_status(self.status_file, self.lock_file, raise_errors=True)
+
+        # Check that a warning is logged
+        expected_log = f"Timed out when trying to read status from '{self.status_file}'"
+        assert expected_log in caplog.text, "Missing expected log message"
+
+    def test_file_not_found_no_fnf_no_errors(self, caplog: "Fixture"):  # noqa: F821
+        """
+        Test the file not found functionality with the `display_fnf_message`
+        and `raise_errors` options both set to False. This should just return
+        an empty dict and not log anything.
+
+        :param caplog: A built-in fixture from the pytest library to capture logs
+        """
+        dummy_file = "i_dont_exist.json"
+        actual_status = read_status(dummy_file, self.lock_file, display_fnf_message=False, raise_errors=False)
+        assert actual_status == {}
+        assert caplog.text == ""
+
+    def test_file_not_found_with_fnf_no_errors(self, caplog: "Fixture"):  # noqa: F821
+        """
+        Test the file not found functionality with the `display_fnf_message`
+        set to True and the `raise_errors` option set to False. This should
+        return an empty dict and log a warning.
+
+        :param caplog: A built-in fixture from the pytest library to capture logs
+        """
+        dummy_file = "i_dont_exist.json"
+        actual_status = read_status(dummy_file, self.lock_file, display_fnf_message=True, raise_errors=False)
+        assert actual_status == {}
+        assert f"Could not find '{dummy_file}'" in caplog.text
+
+    def test_file_not_found_no_fnf_with_errors(self, caplog: "Fixture"):  # noqa: F821
+        """
+        Test the file not found functionality with the `display_fnf_message`
+        set to False and the `raise_errors` option set to True. This should
+        raise a FileNotFound error and not log anything.
+
+        :param caplog: A built-in fixture from the pytest library to capture logs
+        """
+        dummy_file = "i_dont_exist.json"
+        with pytest.raises(FileNotFoundError):
+            read_status(dummy_file, self.lock_file, display_fnf_message=False, raise_errors=True)
+        assert caplog.text == ""
+
+    def test_file_not_found_with_fnf_and_errors(self, caplog: "Fixture"):  # noqa: F821
+        """
+        Test the file not found functionality with the `display_fnf_message`
+        and `raise_errors` options both set to True. This should raise a FileNotFound
+        error and log a warning.
+
+        :param caplog: A built-in fixture from the pytest library to capture logs
+        """
+        dummy_file = "i_dont_exist.json"
+        with pytest.raises(FileNotFoundError):
+            read_status(dummy_file, self.lock_file, display_fnf_message=True, raise_errors=True)
+        assert f"Could not find '{dummy_file}'" in caplog.text
+
+    def test_json_decode_raise_errors_disabled(self, caplog: "Fixture", status_empty_file: str):  # noqa: F821
+        """
+        Test the json decode error functionality with `raise_errors` disabled.
+        This should log a warning and return an empty dict.
+
+        :param caplog: A built-in fixture from the pytest library to capture logs
+        :param status_empty_file: A pytest fixture to give us an empty status file
+        """
+        actual_status = read_status(status_empty_file, self.lock_file, raise_errors=False)
+        assert actual_status == {}
+        assert f"JSONDecodeError raised when trying to read status from '{status_empty_file}'" in caplog.text
+
+    def test_json_decode_raise_errors_enabled(self, caplog: "Fixture", status_empty_file: str):  # noqa: F821
+        """
+        Test the json decode error functionality with `raise_errors` enabled.
+        This should log a warning and raise a JSONDecodeError.
+
+        :param caplog: A built-in fixture from the pytest library to capture logs
+        :param status_empty_file: A pytest fixture to give us an empty status file
+        """
+        with pytest.raises(JSONDecodeError):
+            read_status(status_empty_file, self.lock_file, raise_errors=True)
+        assert f"JSONDecodeError raised when trying to read status from '{status_empty_file}'" in caplog.text
+
+    @pytest.mark.parametrize("exception", [TypeError, ValueError, NotImplementedError, IOError, UnicodeError, OSError])
+    def test_broad_exception_handler_raise_errors_disabled(
+        self, mocker: "Fixture", caplog: "Fixture", exception: Exception  # noqa: F821
+    ):
+        """
+        Test the broad exception handler with `raise_errors` disabled. This should
+        log a warning and return an empty dict.
+
+        :param mocker: A built-in fixture from the pytest-mock library to create a Mock object
+        :param caplog: A built-in fixture from the pytest library to capture logs
+        :param exception: An exception to force `read_status` to raise.
+                          Values for this are obtained from parametrized list above.
+        """
+
+        # Set the mock to raise an exception
+        mock_filelock = mocker.patch("merlin.study.status.FileLock")
+        mock_lock = mocker.MagicMock()
+        mock_lock.acquire.side_effect = exception()
+        mock_filelock.return_value = mock_lock
+
+        actual_status = read_status(self.status_file, self.lock_file, raise_errors=False)
+        assert actual_status == {}
+        assert f"An exception was raised while trying to read status from '{self.status_file}'!" in caplog.text
+
+    @pytest.mark.parametrize("exception", [TypeError, ValueError, NotImplementedError, IOError, UnicodeError, OSError])
+    def test_broad_exception_handler_raise_errors_enabled(
+        self, mocker: "Fixture", caplog: "Fixture", exception: Exception  # noqa: F821
+    ):
+        """
+        Test the broad exception handler with `raise_errors` enabled. This should
+        log a warning and raise whichever exception is passed in (see list of
+        parametrized exceptions in the decorator above).
+
+        :param mocker: A built-in fixture from the pytest-mock library to create a Mock object
+        :param caplog: A built-in fixture from the pytest library to capture logs
+        :param exception: An exception to force `read_status` to raise.
+                          Values for this are obtained from parametrized list above.
+        """
+
+        # Set the mock to raise an exception
+        mock_filelock = mocker.patch("merlin.study.status.FileLock")
+        mock_lock = mocker.MagicMock()
+        mock_lock.acquire.side_effect = exception()
+        mock_filelock.return_value = mock_lock
+
+        with pytest.raises(exception):
+            read_status(self.status_file, self.lock_file, raise_errors=True)
+        assert f"An exception was raised while trying to read status from '{self.status_file}'!" in caplog.text
+
+
+class TestStatusWriting:
+    """Test the logic for writing to status files"""
+
+    status_to_write = {"status": "TESTING"}
+
+    def test_basic_write(self, status_testing_dir: str):
+        """
+        Test the basic functionality of the `write_status` function. This
+        should write status to a file.
+
+        :param status_testing_dir: A pytest fixture defined in `tests/fixtures/status.py`
+                                   that defines a path to the the output directory we'll write to
+        """
+
+        # Test variables
+        status_filepath = f"{status_testing_dir}/basic_write.json"
+        lock_file = f"{status_testing_dir}/basic_write.lock"
+
+        # Run the test
+        write_status(self.status_to_write, status_filepath, lock_file)
+
+        # Check that the path exists and that it contains the dummy status content
+        assert os.path.exists(status_filepath)
+        with open(status_filepath, "r") as sfp:
+            dummy_status = json.load(sfp)
+        assert dummy_status == self.status_to_write
+
+    @pytest.mark.parametrize("exception", [TypeError, ValueError, NotImplementedError, IOError, UnicodeError, OSError])
+    def test_exception_raised(
+        self, mocker: "Fixture", caplog: "Fixture", status_testing_dir: str, exception: Exception  # noqa: F821
+    ):
+        """
+        Test the exception handler using several different exceptions defined in the
+        parametrized list in the decorator above. This should log a warning and not
+        create the status file that we provide.
+
+        :param mocker: A built-in fixture from the pytest-mock library to create a Mock object
+        :param caplog: A built-in fixture from the pytest library to capture logs
+        :param status_testing_dir: A pytest fixture defined in `tests/fixtures/status.py`
+                                   that defines a path to the the output directory we'll write to
+        :param exception: An exception to force `read_status` to raise.
+                          Values for this are obtained from parametrized list above.
+        """
+
+        # Set the mock to raise an exception
+        mock_filelock = mocker.patch("merlin.study.status.FileLock")
+        mock_lock = mocker.MagicMock()
+        mock_lock.acquire.side_effect = exception()
+        mock_filelock.return_value = mock_lock
+
+        # Test variables
+        status_filepath = f"{status_testing_dir}/exception_{exception.__name__}.json"
+        lock_file = f"{status_testing_dir}/exception_{exception.__name__}.lock"
+
+        write_status(self.status_to_write, status_filepath, lock_file)
+        assert f"An exception was raised while trying to write status to '{status_filepath}'!" in caplog.text
+        assert not os.path.exists(status_filepath)
+
+
+class TestStatusConflictHandler:
+    """Test the functionality of the `status_conflict_handler` function."""
+
+    def test_parameter_conflict(self, caplog: "Fixture"):  # noqa: F821
+        """
+        Test that conflicting parameters are handled properly. This is a special
+        case of the use-initial-and-log-warning rule since parameter tokens vary
+        and have to be added to the `merge_rules` dict on the fly.
+
+        :param caplog: A built-in fixture from the pytest library to capture logs
+        """
+
+        # Create two dicts with conflicting parameter values
+        key = "TOKEN"
+        dict_a = {"parameters": {"cmd": {key: "value"}, "restart": None}}
+        dict_b = {"parameters": {"cmd": {key: "new_value"}, "restart": None}}
+        path = ["parameters", "cmd"]
+
+        # Run the test
+        merged_val = status_conflict_handler(
+            dict_a_val=dict_a["parameters"]["cmd"][key], dict_b_val=dict_b["parameters"]["cmd"][key], key=key, path=path
+        )
+
+        # Check that everything ran properly
+        expected_log = (
+            f"Conflict at key '{key}' while merging status files. Defaulting to initial value. "
+            "This could lead to incorrect status information, you may want to re-run in debug mode and "
+            "check the files in the output directory for this task."
+        )
+        assert merged_val == "value"
+        assert expected_log in caplog.text
+
+    def test_non_existent_key(self, caplog: "Fixture"):  # noqa: F821
+        """
+        Test providing `status_conflict_handler` a key that doesn't exist in
+        the `merge_rule` dict. This should log a warning and return None.
+
+        :param caplog: A built-in fixture from the pytest library to capture logs
+        """
+        key = "i_dont_exist"
+        merged_val = status_conflict_handler(key=key)
+        assert merged_val is None
+        assert f"The key '{key}' does not have a merge rule defined. Setting this merge to None." in caplog.text
+
+    def test_rule_string_concatenate(self):
+        """
+        Test the string-concatenate merge rule. This should combine
+        the strings provided in `dict_a_val` and `dict_b_val` into one
+        comma-delimited string.
+        """
+
+        # Create two dicts with conflicting task-queue values
+        key = "task_queue"
+        val1 = "existing_task_queue"
+        val2 = "new_task_queue"
+        dict_a = {key: val1}
+        dict_b = {key: val2}
+
+        # Run the test and make sure the values are being concatenated
+        merged_val = status_conflict_handler(
+            dict_a_val=dict_a[key],
+            dict_b_val=dict_b[key],
+            key=key,
+        )
+        assert merged_val == f"{val1}, {val2}"
+
+    def test_rule_use_initial_and_log_warning(self, caplog: "Fixture"):  # noqa: F821
+        """
+        Test the use-initial-and-log-warning merge rule. This should
+        return the value passed in to `dict_a_val` and log a warning
+        message.
+
+        :param caplog: A built-in fixture from the pytest library to capture logs
+        """
+
+        # Create two dicts with conflicting status values
+        key = "status"
+        dict_a = {key: "SUCCESS"}
+        dict_b = {key: "FAILED"}
+
+        # Run the test
+        merged_val = status_conflict_handler(
+            dict_a_val=dict_a[key],
+            dict_b_val=dict_b[key],
+            key=key,
+        )
+
+        # Check that everything ran properly
+        expected_log = (
+            f"Conflict at key '{key}' while merging status files. Defaulting to initial value. "
+            "This could lead to incorrect status information, you may want to re-run in debug mode and "
+            "check the files in the output directory for this task."
+        )
+        assert merged_val == "SUCCESS"
+        assert expected_log in caplog.text
+
+    def test_rule_use_longest_time_no_dict_a_time(self):
+        """
+        Test the use-longest-time merge rule with no time set for `dict_a_val`.
+        This should default to using the time in `dict_b_val`.
+        """
+        key = "elapsed_time"
+        expected_time = "12h:34m:56s"
+        dict_a = {key: "--:--:--"}
+        dict_b = {key: expected_time}
+
+        merged_val = status_conflict_handler(
+            dict_a_val=dict_a[key],
+            dict_b_val=dict_b[key],
+            key=key,
+        )
+        assert merged_val == expected_time
+
+    def test_rule_use_longest_time_no_dict_b_time(self):
+        """
+        Test the use-longest-time merge rule with no time set for `dict_b_val`.
+        This should default to using the time in `dict_a_val`.
+        """
+        key = "run_time"
+        expected_time = "12h:34m:56s"
+        dict_a = {key: expected_time}
+        dict_b = {key: "--:--:--"}
+
+        merged_val = status_conflict_handler(
+            dict_a_val=dict_a[key],
+            dict_b_val=dict_b[key],
+            key=key,
+        )
+        assert merged_val == expected_time
+
+    def test_rule_use_longest_time(self):
+        """
+        Test the use-longest-time merge rule with times set for both `dict_a_val`
+        and `dict_b_val`. This should use whichever time is longer.
+        """
+
+        # Set up test variables
+        key = "run_time"
+        short_time = "01h:04m:33s"
+        long_time = "12h:34m:56s"
+
+        # Run test with dict b having the longer time
+        dict_a = {key: short_time}
+        dict_b = {key: long_time}
+        merged_val = status_conflict_handler(
+            dict_a_val=dict_a[key],
+            dict_b_val=dict_b[key],
+            key=key,
+        )
+        assert merged_val == "0d:" + long_time  # Time manipulation in status_conflict_handler will prepend '0d:'
+
+        # Run test with dict a having the longer time
+        dict_a_2 = {key: long_time}
+        dict_b_2 = {key: short_time}
+        merged_val_2 = status_conflict_handler(
+            dict_a_val=dict_a_2[key],
+            dict_b_val=dict_b_2[key],
+            key=key,
+        )
+        assert merged_val_2 == "0d:" + long_time
+
+    @pytest.mark.parametrize(
+        "dict_a_val, dict_b_val, expected",
+        [
+            (0, 0, 0),
+            (0, 1, 1),
+            (1, 0, 1),
+            (-1, 0, 0),
+            (0, -1, 0),
+            (23, 20, 23),
+            (17, 21, 21),
+        ],
+    )
+    def test_rule_use_max(self, dict_a_val: int, dict_b_val: int, expected: int):
+        """
+        Test the use-max merge rule. This should take the maximum of 2 values.
+
+        :param dict_a_val: The value to pass in for dict_a_val
+        :param dict_b_val: The value to pass in for dict_b_val
+        :param expected: The expected value from this test
+        """
+        key = "restarts"
+        merged_val = status_conflict_handler(
+            dict_a_val=dict_a_val,
+            dict_b_val=dict_b_val,
+            key=key,
+        )
+        assert merged_val == expected
 
 
 class TestMerlinStatus(unittest.TestCase):